Removed Animated Scrolling from needless places
There is no need for there to be animated scrolling when we manipulate the dom.

he was calling .scrollTo(0,0,1000) which scrolls to the top in 1 second in all of the places where he needed to scroll to the top.

The only place where this makes sense is on the today button so I left it there.  Everywhere else I made the changes So that we scroll imediately (ie 0 seconds)